[geombong/issue-tracker#27] Refactor: move `fetchFilters` function to useDropDown hook
